backend/requirements.txt
aioredis==2.0.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
click==8.1.7
colorlog==6.8.2
exceptiongroup==1.2.0
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
Mako==1.3.2
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
redis==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
